pyproject.toml
[project]
name = "orchard"
version = "2026.1.3"
description = "Python client for Orchard, a compute platform for Apple Silicon"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pillow",
    "huggingface-hub",
    "fastapi[standard]",
    "pydantic-settings",
    "jinja2",
    "sse_starlette",
    "filelock",
    "pynng",
    "pyyaml",
    "requests",
]

[project.scripts]
orchard = "orchard.cli.main:main"

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "httpx",
    "build",
    "twine",
    "ruff",
    "pre-commit",
    "ty>=0.0.1a29",
]

[tool.setuptools.packages.find]
include = [
    "orchard*",
]

[tool.setuptools.package-data]
orchard = [
    "formatter/profiles/**",
]

[dependency-groups]
dev = [
    "pytest",
    "pytest-asyncio",
    "httpx",
    "build",
    "twine",
    "ruff",
    "pre-commit",
    "ty>=0.0.1a29",
]

[tool.ruff.lint]
extend-select = [
    "B",    # flake8-bugbear
    "I",    # isort
    "PGH",  # pygrep-hooks
    "RUF",  # Ruff-specific
    "UP",   # pyupgrade
    "SLF",  # string-literal-format
    "F8",   # flake8-comprehensions
]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 100

[tool.pytest.ini_options]
pythonpath = ["tests"]
